requirements.txt
discord.py>=2.3.2
typing-extensions>=4.7.1
uuid>=1.30
python-dotenv>=1.0.0 